+++ apps/web/package.json
@@ -2,7 +2,6 @@
   "name": "web",
   "version": "0.1.0",
   "private": true,
-  "packageManager": "pnpm@10.28.1+sha512.7d7dbbca9e99447b7c3bf7a73286afaaf6be99251eb9498baefa7d406892f67b879adb3a1d7e687fc4ccc1a388c7175fbaae567a26ab44d1067b54fcb0d6a316",
   "imports": {
     "#src/*": "./src/*"
   },
@@ -11,7 +10,7 @@
     "prebuild": "npm-run-all codegen:pwa-icons codegen:i18n",
     "codegen": "npm-run-all codegen:*",
     "codegen:i18n": "tsx ../../packages/i18n-codegen/generator.ts",
-    "codegen:openapi": "openapi-typescript https://developer.themoviedb.org/openapi/tmdb-api.json -o src/_generated/tmdbV3.d.ts && prettier 'src/_generated/tmdbV3.d.ts' -w",
+    "codegen:openapi": "openapi-typescript https://developer.themoviedb.org/openapi/tmdb-api.json -o ../../packages/tmdb-types/openapi.d.ts && prettier '../../packages/tmdb-types/openapi.d.ts' -w",
     "codegen:pwa-icons": "node ../../packages/pwa-icons/hash-icons.mjs",
     "codegen:tmdb": "node ../../packages/tmdb-codegen/generator.js && prettier 'src/_generated/*.ts' -w",
     "codegen:zod": "tsx ../../packages/tmdb-codegen/generate-selective-zod.ts && prettier 'src/_generated/tmdb-zod.ts' -w",
@@ -21,7 +20,6 @@
     "build:tsc": "tsc --noEmit",
     "start": "next start",
     "lint": "eslint .",
-    "pretest": "pnpm codegen:i18n",
     "test": "vitest",
     "test:ui": "vitest --ui",
     "test:coverage": "vitest --coverage",
@@ -35,6 +33,7 @@
     "@ai-sdk/anthropic": "^3.0.71",
     "@ai-sdk/react": "^3.0.170",
     "@phosphor-icons/react": "^2.1.10",
+    "@repo/tmdb-types": "workspace:*",
     "@serwist/next": "^9.5.7",
     "@stylexjs/stylex": "^0.18.3",
     "@tanstack/react-query": "5.100.1",
@@ -66,11 +65,11 @@
     "@babel/traverse": "^7.29.0",
     "@babel/types": "^7.29.0",
     "@playwright/test": "^1.59.1",
+    "@repo/babel-plugin-stylex-breakpoints": "workspace:*",
+    "@repo/babel-plugin-stylex-css-prop": "workspace:*",
     "@repo/i18n-babel-plugin": "workspace:*",
     "@repo/i18n-codegen": "workspace:*",
     "@repo/pwa-icons": "workspace:*",
-    "@repo/stylex-breakpoints": "workspace:*",
-    "@repo/stylex-css-prop": "workspace:*",
     "@repo/tmdb-codegen": "workspace:*",
     "@repo/tsconfig": "workspace:*",
     "@repo/vector-ingest": "workspace:*",
