+++ packages/babel-plugin-stylex-breakpoints/package.json
@@ -1,14 +1,15 @@
 {
-  "name": "@repo/stylex-breakpoints",
+  "name": "@repo/babel-plugin-stylex-breakpoints",
   "version": "0.0.0",
   "private": true,
   "type": "commonjs",
   "main": "index.js",
+  "scripts": {
+    "lint": "eslint ."
+  },
   "dependencies": {
+    "@babel/core": "7.29.0",
     "@babel/parser": "^7.29.2",
     "@babel/traverse": "^7.29.0"
-  },
-  "peerDependencies": {
-    "@babel/core": "7.29.0"
   }
 }
