+++ packages/eslint-plugin-i18n/package.json
@@ -4,11 +4,14 @@
   "private": true,
   "type": "commonjs",
   "main": "index.js",
-  "peerDependencies": {
+  "scripts": {
+    "lint": "eslint .",
+    "test": "vitest run"
+  },
+  "dependencies": {
     "eslint": "^10.2.1"
   },
   "devDependencies": {
-    "eslint": "^10.2.1",
     "vitest": "^4.1.5"
   }
 }
