+++ packages/i18n-codegen/vitest.config.mts
@@ -1,0 +1,8 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    include: ["**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts}"],
+    exclude: ["**/node_modules/**"],
+  },
+});
