+++ packages/tmdb-types/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "@repo/tsconfig/library.json",
+  "include": ["**/*.ts", "**/*.d.ts"]
+}
